moved dbmgr into graspitCore rather than ivmgr
