+++ include/graspitCore.h
@@ -37,6 +37,10 @@ namespace cmdline{
     class parser;
 }
 
+namespace db_planner {
+    class DatabaseManager;
+}
+
 class MainWindow;
 class IVmgr;
 class TaskDispatcher;
@@ -91,6 +95,9 @@ class GraspitCore
 
   bool headless;
 
+  //! The main and only interface for the CGDB; all interaction with the CGDB should go through this.
+  db_planner::DatabaseManager *mDBMgr;
+
  public:
   GraspitCore();
   ~GraspitCore();
@@ -133,6 +140,15 @@ class GraspitCore
   bool isHeadless(){return headless;}
 
   void emptyWorld();
+
+  //! Get the main database manager, when CGDB support is enabled
+  db_planner::DatabaseManager* getDBMgr(){return mDBMgr;}
+  //! Set the main database manager. Should only be called by the DB connection dialog
+#ifdef CGDB_ENABLED
+  void setDBMgr(db_planner::DatabaseManager *mgr){mDBMgr = mgr;}
+#else
+  void setDBMgr(db_planner::DatabaseManager*){}
+#endif
 
 };
